Drop support for older versions
TODO: find out exactly which versions are now supported
