--- buildsystem-abstraction/adapter-1_12/build.gradle.kts
@@ -1,9 +1,0 @@
-applyAdapterConfiguration()
-
-plugins {
-    `java-library`
-}
-
-dependencies {
-    compileOnly("org.spigotmc:spigot-api:1.12.2-R0.1-SNAPSHOT")
-}
