--- buildsystem-abstraction/adapter-1_13/build.gradle.kts
@@ -1,9 +1,0 @@
-applyAdapterConfiguration()
-
-plugins {
-    `java-library`
-}
-
-dependencies {
-    compileOnly("org.spigotmc:spigot-api:1.13.2-R0.1-SNAPSHOT")
-}
