--- buildsystem-abstraction/adapter-1_14/build.gradle.kts
@@ -1,9 +1,0 @@
-applyAdapterConfiguration()
-
-plugins {
-    `java-library`
-}
-
-dependencies {
-    compileOnly("org.spigotmc:spigot-api:1.14.4-R0.1-SNAPSHOT")
-}
