--- buildsystem-abstraction/adapter-1_17/build.gradle.kts
@@ -1,9 +1,0 @@
-applyAdapterConfiguration()
-
-plugins {
-    `java-library`
-}
-
-dependencies {
-    compileOnly("org.spigotmc:spigot-api:1.17.1-R0.1-SNAPSHOT")
-}
