--- buildsystem-abstraction/adapter-1_20/build.gradle.kts
@@ -1,9 +1,0 @@
-applyAdapterConfiguration()
-
-plugins {
-    `java-library`
-}
-
-dependencies {
-    compileOnly("org.spigotmc:spigot-api:1.20-R0.1-SNAPSHOT")
-}
