--- buildsystem-abstraction/common/build.gradle.kts
@@ -1,11 +1,0 @@
-applyCommonConfiguration()
-
-plugins {
-    `java-library`
-}
-
-dependencies {
-    compileOnly(libs.spigot)
-    compileOnly(libs.paperlib)
-    compileOnly(libs.annotations)
-}
