--- buildsystem-abstraction/common/src/main/java/de/eintosti/buildsystem/version/customblocks/CustomBlocks.java
@@ -1,39 +1,0 @@
-/*
- * Copyright (c) 2018-2025, Thomas Meaney
- * Copyright (c) contributors
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-package de.eintosti.buildsystem.version.customblocks;
-
-import org.bukkit.block.Block;
-import org.bukkit.block.BlockFace;
-import org.bukkit.entity.Player;
-import org.bukkit.event.block.BlockPlaceEvent;
-import org.bukkit.event.player.PlayerInteractEvent;
-
-public interface CustomBlocks {
-
-    void setBlock(BlockPlaceEvent event, String key);
-
-    void setPlant(PlayerInteractEvent event);
-
-    void modifySlab(PlayerInteractEvent event);
-
-    void toggleIronTrapdoor(PlayerInteractEvent event);
-
-    void toggleIronDoor(PlayerInteractEvent event);
-
-    void rotateBlock(Block block, Player player, BlockFace direction);
-}
