+++ buildsystem-core/src/main/java/de/eintosti/buildsystem/listener/EntitySpawnListener.java
@@ -20,9 +20,7 @@ package de.eintosti.buildsystem.listener;
 import de.eintosti.buildsystem.BuildSystemPlugin;
 import de.eintosti.buildsystem.api.world.BuildWorld;
 import de.eintosti.buildsystem.storage.WorldStorageImpl;
-import de.eintosti.buildsystem.util.EntityAIManager;
 import org.bukkit.World;
-import org.bukkit.entity.Entity;
 import org.bukkit.entity.LivingEntity;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
@@ -49,9 +47,8 @@ public class EntitySpawnListener implements Listener {
             return;
         }
 
-        Entity entity = event.getEntity();
-        if (entity instanceof LivingEntity) {
-            EntityAIManager.setAIEnabled((LivingEntity) entity, false);
+        if (event.getEntity() instanceof LivingEntity livingEntity) {
+            livingEntity.setAI(false);
         }
     }
 }
