+++ buildsystem-core/src/main/java/de/eintosti/buildsystem/player/settings/SettingsInventory.java
@@ -27,7 +27,6 @@ import de.eintosti.buildsystem.api.player.settings.DesignColor;
 import de.eintosti.buildsystem.api.player.settings.Settings;
 import de.eintosti.buildsystem.config.ConfigValues;
 import de.eintosti.buildsystem.util.InventoryUtils;
-import de.eintosti.buildsystem.version.util.MinecraftVersion;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
@@ -75,16 +74,10 @@ public class SettingsInventory implements Listener {
         addSettingsItem(player, inventory, 30, XMaterial.PAPER, settings.isScoreboard(),
                 configValues.isScoreboard() ? "settings_scoreboard_item" : "settings_scoreboard_disabled_item",
                 configValues.isScoreboard() ? "settings_scoreboard_lore" : "settings_scoreboard_disabled_lore");
-        addSettingsItem(player, inventory, 31, getSlabBreakingMaterial(), settings.isSlabBreaking(), "settings_slab_breaking_item", "settings_slab_breaking_lore");
+        addSettingsItem(player, inventory, 31, XMaterial.SMOOTH_STONE_SLAB, settings.isSlabBreaking(), "settings_slab_breaking_item", "settings_slab_breaking_lore");
         addSettingsItem(player, inventory, 32, XMaterial.MAGMA_CREAM, settings.isSpawnTeleport(), "settings_spawnteleport_item", "settings_spawnteleport_lore");
 
         return inventory;
-    }
-
-    private XMaterial getSlabBreakingMaterial() {
-        return MinecraftVersion.getCurrent().isEqualOrHigherThan(MinecraftVersion.AQUATIC_13)
-                ? XMaterial.SMOOTH_STONE_SLAB
-                : XMaterial.STONE_SLAB;
     }
 
     public void openInventory(Player player) {
